Migrates general decryption to DecryptEnvironmentPostProcessor
This is only used if legacy bootstrap is not being used.

This works for both @ConfigurationProperties and calls to Environment.

TextEncryptorBindHandler is only used in ConfigData Bootstrapper.

Existing EnvironmentDecryptApplicationInitializer is only used if legacy bootstrap is used.

Some code could be put in an abstract class that EnvironmentDecryptApplicationInitializer and DecryptEnvironmentPostProcessor could share.